# docs/openapi/pn-radd-bff-internal.yaml
openapi: 3.0.1
info:
  title: 'Piattaforma Notifiche: API del BFF dei servizi RADD'
  description: |
    API utilizzate dal Backend for Frontend dei servizi RADD.
  version: '1.0'
servers:
  - url: https://tbd/
...
paths:
...
  /radd-web/aor/transaction/start:
    parameters:
      - $ref: 'remote-refs.yaml#/components/parameters/uidAuthFleet'          # NO EXTERNAL
    post:
      tags:
        - aorTransactionManagement
      operationId: startAorTransaction
      description: API utilizzata per la richiesta di avvio delloperazione di scaricamento e consegna degli avvisi di avvenuta ricezione.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AorStartTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/aor/transaction/complete:
    parameters:
      - $ref: 'remote-refs.yaml#/components/parameters/uidAuthFleet'          # NO EXTERNAL
    post:
      tags:
        - aorTransactionManagement
      operationId: completeAorTransaction
      description: API utilizzata per la notifica del completamento dell'operazione di consegna dell'atto.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/aor/transaction/abort:
    parameters:
      - $ref: 'remote-refs.yaml#/components/parameters/uidAuthFleet'          # NO EXTERNAL
    post:
      tags:
        - aorTransactionManagement
      operationId: abortAorTransaction
      description: API utilizzata per la notifica dell'annullamento dell'operazione di consegna del/degli avviso/i di avvenuta ricezione.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/act/operations/by-id/{operationId}:
    parameters:
      - $ref: '#/components/parameters/operationId'
    get:
      tags:
        - notificationInquiry
      operationId: getActTransactionByOperationId
      description: API utilizzata per il recupero di una transazione legata ad un atto opponibille ai terzi.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationActResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/aor/operations/by-id/{operationId}:
    parameters:
      - $ref: '#/components/parameters/operationId'
    get:
      tags:
        - notificationInquiry
      operationId: getAorTransactionByOperationId
      description: API utilizzata per il recupero di una transazione legata ad un AAR.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAorResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/act/operations/by-iun/{iun}:
    parameters:
      - $ref: '#/components/parameters/iun'
    get:
      tags:
        - notificationInquiry
      operationId: getActPracticesByIun
      description: API utilizzata per il recupero di tutte le pratiche legate ad atti opponibili ai terzi ed associate ad uno specifico iun.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/aor/operations/by-iun/{iun}:
    parameters:
      - $ref: '#/components/parameters/iun'
    get:
      tags:
        - notificationInquiry
      operationId: getAorPracticesByIun
      description: API utilizzata per il recupero di tutte le pratiche legate ad un AAR ed associate ad uno specifico iun.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/act/operations/by-internalId/{taxId}:
    parameters:
      - $ref: '#/components/parameters/taxId'
    post:
      tags:
        - notificationInquiry
      operationId: getActPracticesByInternalId
      description: API utilizzata per il recupero di tutte le pratiche legate ad atti opponibili ai terzi ed associate ad uno specifico internal id.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsActDetailsResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-web/aor/operations/by-internalId/{taxId}:
    parameters:
      - $ref: '#/components/parameters/taxId'
    post:
      tags:
        - notificationInquiry
      operationId: getAorPracticesByInternalId
      description: API utilizzata per il recupero di tutte le pratiche legate ad AAR ed associate ad uno specifico internal id.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsAorDetailsResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
components:
  schemas:
    FilterRequest:
      description: Campi per poter applicare filtri sui dati a DB
      type: object
      properties:
        from:
          type: string
          format: date-time
          description: Data/ora in formato RFC3339.
        to:
          type: string
          format: date-time
          description: Data/ora in formato RFC3339.
...
    StartTransactionResponse:
      description: Response alla richiesta di avvio della transazione
      type: object
      properties:
        urlList:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/StartTransactionResponseStatus'
    AorStartTransactionRequest:
      description: Richiesta  di completamento operazione
      type: object
      required:
        - versionToken
        - fileKey
        - operationId
        - recipientTaxId
        - checksum
      properties:
        versionToken:
          type: string
        fileKey:
          type: string
        operationId:
          type: string
        recipientTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "TNTGTR76E21H751S"
        recipientType:
          type: string
          enum: ["PF", "PG"]
          example: "PF"
        delegateTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "RFRGRZ66E21H751B"
        operationDate:
          type: string
          format: date-time
          description: Data/ora di completamento della transazione,  espressa in formato RFC3339.
          example: "2022-06-21T11:44:28Z"
        checksum:
          type: string
          example: "YTlkZGRkNzgyZWM0NzkyODdjNmQ0NGE5ZDM2YTg4ZjQ5OTE1ZGM2NjliYjgzNzViMTZhMmE5ZmE3NmE4ZDQzNwo="
          description: Checksum SHA256 del file caricato nella chiamata di documetUpload, esepresso in base64
    CompleteTransactionRequest:
      description: Richiesta  di chiusura della transazione
      type: object
      required:
        - operationId
        - operationDate
      properties:
        operationId:
          type: string
          example: "2bb5c29e0cf411ed861d0242ac120002"
        operationDate:
          type: string
          format: date-time
          description: Data/ora di completamento della transazione,  espressa in formato RFC3339.
          example: "2022-06-21T11:44:28Z"
    CompleteTransactionResponse:
      description: Response alla richiesta di chiusura della transazione
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TransactionResponseStatus'
    AbortTransactionRequest:
      description: Richiesta  di cancellazione della transazione
      type: object
      required:
        - operationId
        - operationDate
      properties:
        operationId:
          type: string
          example: "2bb5c29e0cf411ed861d0242ac120002"
        reason:
          type: string
          default: "abort"
        operationDate:
          type: string
          format: date-time
          description: Data/ora di annullamento della transazione,  espressa in formato RFC3339.
          example: "2022-06-21T11:44:28Z"
    AbortTransactionResponse:
      description: Response alla richiesta di chiusura della transazione
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TransactionResponseStatus'
    ResponseStatus:
      description: Stato della response
      type: object
      properties:
        code:
          description: |
            Return code.
            * 0 -> OK
            * 99 -> KO
          type: number
          enum: [0, 99]
        message:
          type: string
          example: OK
    StartTransactionResponseStatus:
      description: Stato della response
      allOf:
        - $ref: '#/components/schemas/ResponseStatus'
        - type: object
          properties:
            code:
              description: |
                Return code.
                * 0 -> OK
                * 99 -> KO
                * 2 -> NotReady
              type: number
              enum: [0, 99, 2]
            retryAfter:
              description: Tempo, espresso in ms comunicato  al chiamante, trascorso il quale è possibile effettuare un nuovo tentativo di avvio della transazione.
              type: number
              example: 670
    TransactionResponseStatus:
      description: Stato della response
      allOf:
        - $ref: '#/components/schemas/ResponseStatus'
        - type: object
          properties:
            code:
              description: |
                Return code.
                * 0 -> OK (Anche in caso di  chiamata successiva con identici parametri)
                * 1 -> Transazione inesistente
                * 2 -> Transazione già completata (Messaggio di errore differente se già completata per chiamata a complete piuttosto che abort)
                * 99 -> KO
              type: number
              enum: [0,1,2,99]
...
    OperationsResponse:
      description: Pratiche recuperate
      type: object
      properties:
        operations:
          type: array
          items:
            $ref: '#/components/schemas/OperationsDetailsResponse'
        result:
          type: boolean
        status:
          $ref: '#/components/schemas/OperationResponseStatus'
    OperationsDetailsResponse:
      description: Dettagli pratiche recuperate
      type: object
      properties:
        operationStatus:
          type: string
          example: "COMPLETED"
        operationType:
          type: string
          example: "ACT"
        iuns:
          type: array
          items:
            type: string
            example: "welcome"
        operationId:
          type: string
          example: "asd1234"
        fileKey:
          type: string
          example: "FL-23-KKL"
        qrCode:
          type: string
          example: "FBD"
        recipientTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "TNTGTR76E21H751S"
        recipientType:
          type: string
          example: "PF"
        delegateTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "RFRGRZ66E21H751B"
        uid:
          type: string
          example: "uddHHHw123mk"
        operationStartDate:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
        operationEndDate:
          type: string
          format: date-time
          example: 2017-07-21T18:00:28Z
        errorReason:
          type: string
          example: "Error"
    OperationsActDetailsResponse:
      description: Id pratiche recuperate
      type: object
      properties:
        elements:
          type: array
          items:
            $ref: '#/components/schemas/OperationActDetailResponse'
        result:
          type: boolean
        status:
          $ref: '#/components/schemas/OperationResponseStatus'
    OperationsAorDetailsResponse:
      description: Id pratiche recuperate
      type: object
      properties:
        elements:
          type: array
          items:
            $ref: '#/components/schemas/OperationAorDetailResponse'
        result:
          type: boolean
        status:
          $ref: '#/components/schemas/OperationResponseStatus'
    OperationResponseStatus:
      description: Stato della response
      allOf:
        - $ref: '#/components/schemas/ResponseStatus'
        - type: object
          properties:
            code:
              description: |
                Return code.
                * 0 -> OK
                * 1 -> Nessun elemento associato
                * 99 -> KO
              type: number
              enum: [0,1,99]
    OperationActResponse:
      description: Risposta della notifica
      type: object
      properties:
        element:
          $ref: '#/components/schemas/OperationActDetailResponse'
        result:
          type: boolean
        status:
          $ref: '#/components/schemas/OperationResponseStatus'
    OperationActDetailResponse:
      description: Dettagli della response Act
      type: object
      properties:
        operationStatus:
          type: string
          example: "COMPLETED"
        operationType:
          type: string
          example: "ACT"
        iun:
          type: string
          example: "LJLH-GNTJ-DVXR-202209-J-1"
        operationId:
          type: string
          example: "asd1234"
        fileKey:
          type: string
          example: "FL-23-KKL"
        qrCode:
          type: string
          example: "FBD"
        recipientTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "TNTGTR76E21H751S"
        recipientType:
          type: string
          example: "PF"
        delegateTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "RFRGRZ66E21H751B"
        uid:
          type: string
          example: "uddHHHw123mk"
        operationStartDate:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
        operationEndDate:
          type: string
          format: date-time
          example: 2017-07-21T18:00:28Z
        errorReason:
          type: string
          example: "Error"
    OperationAorResponse:
      description: Risposta della notifica
      type: object
      properties:
        element:
          $ref: '#/components/schemas/OperationAorDetailResponse'
        result:
          type: boolean
        status:
          $ref: '#/components/schemas/OperationResponseStatus'
    OperationAorDetailResponse:
      description: Dettagli della response Act
      type: object
      properties:
        operationStatus:
          type: string
          example: "COMPLETED"
        operationType:
          type: string
          example: "ACT"
        iuns:
          type: array
          items:
            type: string
            example: "welcome"
        operationId:
          type: string
          example: "asd1234"
        fileKey:
          type: string
          example: "FL-23-KKL"
        qrCode:
          type: string
          example: "FBD"
        recipientTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "TNTGTR76E21H751S"
        recipientType:
          type: string
          example: "PF"
        delegateTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "RFRGRZ66E21H751B"
        uid:
          type: string
          example: "uddHHHw123mk"
        operationStartDate:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
        operationEndDate:
          type: string
          format: date-time
          example: 2017-07-21T18:00:28Z
        errorReason:
          type: string
          example: "Error"
  parameters:
...
    taxId:
      x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
      name: taxId
      description: Codice Fiscale
      in: path
      required: true
      schema:
        type: string
      example: TNTGTR76E21H751S
...
    operationId:
      name: operationId
      description: Id della pratica
      in: path
      required: true
      schema:
        type: string
      example: "asd123"
...
    iun:
      name: iun
      description: Identification code
      in: path
      required: true
      schema:
        type: string
      example: "LJLH-GNTJ-DVXR-202209-J-1"
...
  responses:
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: 'remote-refs.yaml#/components/schemas/Problem'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: 'remote-refs.yaml#/components/schemas/Problem'
    Invalid:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: 'remote-refs.yaml#/components/schemas/Problem'
    NotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: 'remote-refs.yaml#/components/schemas/Problem'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: 'remote-refs.yaml#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: 'remote-refs.yaml#/components/schemas/Problem'


#  securitySchemes:                                                             # ONLY EXTERNAL
#    APIKeyAuth:                                                                # ONLY EXTERNAL
#      type: apiKey                                                             # ONLY EXTERNAL
#      in: header                                                               # ONLY EXTERNAL
#      name: x-api-key                                                          # ONLY EXTERNAL

#security:                                                                      # ONLY EXTERNAL
#  - APIKeyAuth: []                                                             # ONLY EXTERNAL